Reorganize indexing to be implemented inside a mix-in, to allow easier integration into both draft and split store
