+++ common/lib/xmodule/xmodule/modulestore/split_mongo/split.py
@@ -78,6 +78,7 @@ from xmodule.modulestore import (
 
 from ..exceptions import ItemNotFoundError
 from .caching_descriptor_system import CachingDescriptorSystem
+from xmodule.modulestore.courseware_index import ModuleStoreCoursewareIndexMixin
 from xmodule.modulestore.split_mongo.mongo_connection import MongoConnection, DuplicateKeyError
 from xmodule.modulestore.split_mongo import BlockKey, CourseEnvelope
 from xmodule.error_module import ErrorDescriptor
@@ -591,7 +592,7 @@ class SplitBulkWriteMixin(BulkOperationsMixin):
         return structures
 
 
-class SplitMongoModuleStore(SplitBulkWriteMixin, ModuleStoreWriteBase):
+class SplitMongoModuleStore(SplitBulkWriteMixin, ModuleStoreWriteBase, ModuleStoreCoursewareIndexMixin):
     """
     A Mongodb backed ModuleStore supporting versions, inheritance,
     and sharing.
